秒杀商品

+++ base_resource/src/main/resources/database/master.sql
@@ -666,7 +666,7 @@ ADD COLUMN `template`  varchar(255) NULL AFTER `show`,
 ADD COLUMN `create_datetime`  datetime NULL AFTER `template`;
 
 ALTER TABLE `product_picture`
-ADD COLUMN `sort`  integer(255) NULL AFTER `product_model_id`;
+ADD COLUMN `sort`  integer(16) NULL AFTER `product_model_id`;
 
 ALTER TABLE `product`
 ADD COLUMN `show_datetime`  datetime NULL AFTER `fans_amount`;
